Opening all dialogs as sheets to ensure visibility
On drop the topmost window is the source window of the dragged object, not DropPy. This means an error dialog may pop up somewhere bwlow (in the center of the screen). This way bad from a usability standpoint, just the icon was bouncing in the dock and progress seemed halted in DropPy’s main window.

Signed-off-by: Günther Eberl <guenther@eberl.se>